ajout box

+++ Apprentis_Recycleur/src/objets/Trash.java
@@ -24,7 +24,7 @@ public class Trash implements Selectionnable, Dessinable {
 		this.nom = nomTrash;
 		this.idType = id;
 		
-		point = new Point2D.Double(500, 600);
+		point = new Point2D.Double(1070, 200);
 	}
 	
 	public Trash(String nomTrash, Systems.TypeDechet id, String image) {
